+++ src/prelude.rs
@@ -6,10 +6,10 @@ pub use display_interface_spi::{SPIInterface, SPIInterfaceNoCS};
 
 pub use super::{
     brightness::Brightness,
-    displayrotation::DisplayRotation,
-    displaysize::{
-        DisplaySize128x32, DisplaySize128x64, DisplaySize64x48, DisplaySize72x40, DisplaySize96x16,
+    mode::DisplayConfig,
+    rotation::DisplayRotation,
+    size::{
+        DisplaySize, DisplaySize128x32, DisplaySize128x64, DisplaySize64x48, DisplaySize72x40,
+        DisplaySize96x16,
     },
-    mode::{displaymode::DisplayModeTrait, GraphicsMode, TerminalMode},
-    properties::DisplayProperties,
 };