+++ src/size.rs
@@ -1,4 +1,4 @@
-//! Display size
+//! Display size.
 
 use super::command::Command;
 use display_interface::{DisplayError, WriteOnlyDataCommand};
@@ -7,7 +7,7 @@ use generic_array::{
     ArrayLength,
 };
 
-/// Display information
+/// Display information.
 ///
 /// This trait describes information related to a particular display.
 /// This includes resolution, offset and framebuffer size.
@@ -36,8 +36,8 @@ pub trait DisplaySize {
 
     /// Send resolution and model-dependent configuration to the display
     ///
-    /// See [`Command::ComPinConfig`](../command/enum.Command.html#variant.ComPinConfig)
-    /// and [`Command::InternalIref`](../command/enum.Command.html#variant.InternalIref)
+    /// See [`Command::ComPinConfig`](crate::Command::ComPinConfig)
+    /// and [`Command::InternalIref`](crate::Command::InternalIref)
     /// for more information
     fn configure(&self, iface: &mut impl WriteOnlyDataCommand) -> Result<(), DisplayError>;
 }